requirements.txt
fastapi
uvicorn
pandas
python-dotenv
langchain
langchain-community
langchain-core
langchain-anthropic
langgraph
pinecone-client
sentence-transformers
openai
snowflake-connector-python
matplotlib
serpapi
graphviz
langchain_openai
ipykernel
pyarrow<19.0.0
boto3
bs4
sqlachemy
snowflake-sqlalchemy